Move traversal test models
